refactor(solana payments): clean up file structure into files per token PE-5993
